fix: certified output for competition 2015
Nasty hack to fix the INPUT_MISMATCH problem for the certified output in
the competition. The problem is that the ceta-postproc only checks
syntactic equality. The equality check may fail as we use a set
representation for TRSs. We can not do what is done in tct2 (parser
returns (rulelist,problem)) as we have use a more generic main function
that depends on the TRS mode.

So, we store the original rules in the problem type and use the
xml/toCeTA instance of the problem to differentiate output for initial
and sub-problems.